In proxy, rewrite /etc/hosts from 'weave attach', via new utility weavehosts
